+++ src/Api/Service/RevisionSaver.php
@@ -76,8 +76,11 @@ class RevisionSaver {
 		if ( $entityId !== null ) {
 			$params['id'] = $entityId->getSerialization();
 
-			// Always clear so that removing elements is possible
-			$params['clear'] = 'true';
+			// If we are provided an empty entity, then set the clear flag
+			if ( $entity->isEmpty() ) {
+				$params['clear'] = true;
+			}
+
 			// Add more detail to the default "Cleared an entity" summary
 			// Note: this is later overridden if a summary is provided in the EditInfo
 			$params['summary'] = 'Edited a ' . $entity->getType();